Rename method, update doc headers
