+++ codecov.yml
@@ -7,10 +7,12 @@ coverage:
   status:
     patch:
       default:
+        threshold: 50
         if_no_uploads: error
     changes: true
     project:
       default:
         target: auto
+        threshold: 2
         if_no_uploads: error
 comment: false
